Update assemblePartials to store partials with base directory name
I plan to use Assemble to create a pattern library, using Brad Frost's atomic design principles as a foundation. Put simply, I just need to change the way partials are stored, so that we can call them as follows:

```handlebars
{{> directory-filename}}
```

where in fact, the partial is physically stored in

```
directory/(subdirectory/)*filename.hbs
```

In order to do that, I need to update the way partials are registered in Assemble, as follows:

(1) Add an option `partialsdir` akin to `options.layoutdir`
(2) Update `assemblePartials` to store partials into `assemble.options.data[directory + '-' + filename]`

And that's all, hopefully.